requirements.txt
pandas
SQLAlchemy
psycopg2-binary
python-dotenv
requests
beautifulsoup4